requirements.txt
requests
databricks-sql-connector
fuzzywuzzy
python-Levenshtein